Updated some of the code that generates the grades table
The table on line 101 can be removed, but I have left it because the code I have is incomplete. I am still doing a commit because I want to at least have a version of the code that will display the table of the assignment grades that has a dedicated column for the median as well (the output for the column of the medians is what I have been trying to fix with little success). This way, in the event that I mess up something, I have a version that I can go back and see.